+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Fixed a bug where all checkboxes would appear checked within selected Matrix blocks. 
 - Fixed a bug where Recent Entries widgets were getting mangled when new entries were created via Quick Post widgets.
 - Fixed a bug where linked chip labels weren’t getting focus rings.
+- Fixed an error that could occur when field handles were overridden. ([#14166](https://github.com/craftcms/cms/issues/14166), [#14221](https://github.com/craftcms/cms/issues/14221))
 
 ## 5.0.0-alpha.8 - 2024-01-23